# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cyber Code API",
    "version": "1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4747"
    }
  ],
  "security": [
    {
      "basic": []
    }
  ],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "Data_LocationInfo": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LocationInfo"
          }
        },
        "required": [
          "data"
        ]
      },
      "LocationInfo": {
        "type": "object",
        "properties": {
          "directory": {
            "type": "string"
          },
          "workspace": {
            "type": [
              "string",
              "null"
            ]
          },
          "project": {
            "$ref": "#/components/schemas/ProjectInfo"
          }
        },
        "required": [
          "directory",
          "project"
        ]
      },
      "ProjectInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "directory": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "directory"
        ]
      },
...
      "RetrySafety": {
        "type": "string",
        "enum": [
          "read_only",
          "idempotent",
          "reconcile",
          "never"
        ]
      },
...
      "Data_Array_of_PendingRequest": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PendingRequest"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "PendingRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "call_id": {
            "type": "string"
          },
          "message_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "session_id",
          "call_id",
          "message_id"
        ],
        "oneOf": [
          {
            "$ref": "#/components/schemas/PermissionAsk",
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "permission"
              }
            },
            "required": [
              "kind"
            ]
          },
          {
            "type": "object",
            "properties": {
              "questions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Question"
                }
              },
              "kind": {
                "type": "string",
                "const": "question"
              }
            },
            "required": [
              "kind",
              "questions"
            ]
          }
        ],
        "description": "A request waiting for a client, as listed to clients."
      },
      "PermissionAsk": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "resources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "always_patterns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Patterns saved by an `always` reply."
          },
          "metadata": {
            "description": "Display data such as a diff or the command.",
            "default": null
          }
        },
        "required": [
          "action",
          "resources",
          "always_patterns"
        ]
      },
      "Question": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "header": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionOption"
            }
          },
          "multi_select": {
            "type": "boolean",
            "default": false
          },
          "allow_custom": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "question",
          "header",
          "options"
        ]
      },
      "QuestionOption": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "default": ""
          }
        },
        "required": [
          "label"
        ]
      },
      "Located_Array_of_ModelInfo": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            }
          }
        },
        "required": [
          "location",
          "data"
        ],
        "description": "A Location-scoped response."
      },
      "ModelInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "`provider/model`."
          },
          "provider": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "available": {
            "type": "boolean"
          },
          "context_limit": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "reasoning": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "provider",
          "name",
          "available",
          "context_limit",
          "reasoning"
        ],
        "description": "A model offered to clients."
      },
      "Located_Array_of_AgentInfo": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentInfo"
            }
          }
        },
        "required": [
          "location",
          "data"
        ],
        "description": "A Location-scoped response."
      },
      "AgentInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "description": "`primary`, `subagent` or `all`."
          }
        },
        "required": [
          "name",
          "description",
          "mode"
        ],
        "description": "An agent clients can select."
      },
      "Located_Array_of_ToolInfo": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolInfo"
            }
          }
        },
        "required": [
          "location",
          "data"
        ],
        "description": "A Location-scoped response."
      },
      "ToolInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "input_schema": true,
          "retry_safety": {
            "$ref": "#/components/schemas/RetrySafety",
            "description": "`read_only`, `idempotent`, `reconcile` or `never`."
          },
          "concurrency_safe": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "description",
          "input_schema",
          "retry_safety",
          "concurrency_safe"
        ]
      },
      "Located_Array_of_CommandInfo": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommandInfo"
            }
          }
        },
        "required": [
          "location",
          "data"
        ],
        "description": "A Location-scoped response."
      },
      "CommandInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "description": "`builtin`, `skill` or `command`."
          },
          "argument_hint": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "name",
          "description",
          "source"
        ],
        "description": "A slash command or skill for autocomplete."
      },
      "Located_Array_of_string": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "location",
          "data"
        ],
        "description": "A Location-scoped response."
      },
      "ErrorBody": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "title": "ErrorBody",
        "description": "The JSON body of every declared failure.",
        "type": "object",
        "properties": {
          "_tag": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "ref": {
            "description": "`err_<id>` for `UnknownError`, logged server-side.",
            "type": [
              "string",
              "null"
            ]
          },
          "paths": {
            "description": "Paths for rewind conflicts.",
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "_tag",
          "message"
        ]
      }
    }
  },
  "paths": {
...
    "/api/v1/openapi.json": {
      "get": {
        "operationId": "v1.health.openapi",
        "summary": "This document",
        "tags": [
          "health"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/location": {
      "get": {
        "operationId": "v1.location.get",
        "summary": "Resolve a Location",
        "tags": [
          "location"
        ],
        "parameters": [
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_LocationInfo"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/permissions/requests": {
      "get": {
        "operationId": "v1.permission.list",
        "summary": "Pending permission requests",
        "tags": [
          "permission"
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Array_of_PendingRequest"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/requests": {
      "get": {
        "operationId": "v1.question.list",
        "summary": "Pending questions",
        "tags": [
          "question"
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Array_of_PendingRequest"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models": {
      "get": {
        "operationId": "v1.model.list",
        "summary": "Models, available first",
        "tags": [
          "model"
        ],
        "parameters": [
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Located_Array_of_ModelInfo"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents": {
      "get": {
        "operationId": "v1.agent.list",
        "summary": "Selectable agents",
        "tags": [
          "agent"
        ],
        "parameters": [
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Located_Array_of_AgentInfo"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools": {
      "get": {
        "operationId": "v1.tool.list",
        "summary": "Tools offered for an agent and mode",
        "tags": [
          "tool"
        ],
        "parameters": [
          {
            "name": "agent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Located_Array_of_ToolInfo"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/schema": {
      "get": {
        "operationId": "v1.tool.schema",
        "summary": "Built-in tool schema bundle",
        "tags": [
          "tool"
        ],
        "parameters": [
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Located_Array_of_ToolInfo"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commands": {
      "get": {
        "operationId": "v1.command.list",
        "summary": "Commands and skills for autocomplete",
        "tags": [
          "command"
        ],
        "parameters": [
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Located_Array_of_CommandInfo"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fs/find": {
      "get": {
        "operationId": "v1.fs.find",
        "summary": "Find files for @ mentions",
        "tags": [
          "fs"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Located_Array_of_string"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  }
}